fix: fix where vec2 would convert to float instead of double

+++ src/CutieDesigner/datatypes/Vec2Data.cpp
@@ -1,9 +1,9 @@
 #include "Vec2Data.hpp"
-#include "DecimalData.hpp"
+#include <qvectornd.h>
 
 Vec2DataType::Vec2DataType() : NodeDataType("vec2-01", "Vec2", {"vec2-01", "decimal"}) {}
 
 Vec2Data::Vec2Data(const QVector2D &v) {
   registerConvert([&v]() { return v; });
-  registerConvert([&v]() { return v.length(); });
+  registerConvert([&v]() { return (double)v.length(); });
 }
